ClientInstanceStore
