+++ lib/db/client-instance-store.js
@@ -4,26 +4,9 @@
 const metricsHelper = require('../metrics-helper');
 const { DB_TIME } = require('../events');
 
-const COLUMNS = [
-    'app_name',
-    'instance_id',
-    'sdk_version',
-    'client_ip',
-    'last_seen',
-    'created_at',
-];
-const TABLE = 'client_instances';
-
-const ONE_DAY = 24 * 61 * 60 * 1000;
-
-const mapRow = row => ({
-    appName: row.app_name,
-    instanceId: row.instance_id,
-    sdkVersion: row.sdk_version,
-    clientIp: row.client_ip,
-    lastSeen: row.last_seen,
-    createdAt: row.created_at,
-});
+const { ONE_DAY } = require('./utils/const/timings');
+const { TABLE, COLUMNS } = require('./utils/const/instance-store');
+const { mapRow } = require('./utils/mappings/client-instance-store');
 
 class ClientInstanceStore {
     constructor(db, eventBus, getLogger) {
