+++ lib/db/utils/const/timings.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const ONE_DAY = 24 * 61 * 60 * 1000;
+
+module.exports.ONE_DAY = ONE_DAY;
